+++ graphics/tuple.go
@@ -5,5 +5,5 @@ type Tuple struct {
 	X float32
 	Y float32
 	Z float32
-	W int
+	W float32
 }
